wrapper: fix segfault in cgroup_get_value_int64()
The second and third arguments passed to cgroup_get_value_int64() are of
type char * and the user might pass NULL in place of one or both of the
arguments, causing a segfault. The reason is, argument values are used
without checks, fix it by checking for NULL pointers before proceeding.

Reproducer:
-----------
 #include <stdlib.h>
 #include <libcgroup.h>

 int main(void)
 {
 	struct cgroup_controller *cgc;
 	struct cgroup *cgrp;
 	int ret;

 	ret = cgroup_init();
 	if (ret)
 		exit(1);

 	cgrp = cgroup_new_cgroup("fuzzer");
 	if (!cgrp)
 		exit(1);

 	cgc = cgroup_add_controller(cgrp, "cpu");
 	if (!cgc)
 		exit(1);

 	ret = cgroup_add_value_string(cgc, "cpu.shares", "512");
 	if (ret)
 		exit (1);

 	cgroup_get_value_int64(cgc, NULL, NULL);
 	// should not reach here.
 	return 0;
 }

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>