wrapper: Make cgroup_compare_cgroup() order agnostic
cgroup_compare_cgroup() can be used to compare two struct cgroup
instances.  Improve its comparison algorithm so that it can successfully
identify cgroups as the same if the only difference is the order of the
controllers stored within them.

With this change, the following two cgroups would be identified as
equal.

cgroup_a
	name = foo
	controller[0] = memory
	controller[1] = cpu

cgroup_b
	name = foo
	controller[0] = cpu
	controller[1] = memory

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>