wrapper: fix segfault in cgroup_get_value_bool()
The second and third arguments passed to cgroup_get_value_bool() are of
type pointers and the user might pass NULL in place of one or both of
the arguments, causing a segfault. The reason is, argument values are
used without checks, fix it by checking for NULL pointers before
proceeding.

Reproducer:
-----------

 #include <stdlib.h>
 #include <libcgroup.h>

 int main(void)
 {
 	struct cgroup_controller *cgc;
 	struct cgroup *cgrp;
 	int ret;

 	ret = cgroup_init();
 	if (ret)
 		exit(1);

 	cgrp = cgroup_new_cgroup("fuzzer");
 	if (!cgrp)
 		exit(1);

 	cgc = cgroup_add_controller(cgrp, "cpuset");
 	if (!cgc)
 		exit(1);

 	ret = cgroup_add_value_string(cgc, "cpuset.cpu_exclusive", "0");
 	if (ret)
 		exit (1);

 	cgroup_get_value_bool(cgc, "cpuset.cpu_exclusive", NULL);
	//should not reach here
 	return 0;
 }

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>