Revert "api: cgroup v2 – enable controller at the leaf node"
This reverts commit bb291add5db87fe36a82a18c6eb6aaf64dc07694.

cgroup v2, has no internal process constraint, where the process runs
only on the leaf node of the cgroup hierarchy and no controllers should
be enabled on the leaf cgroup node too, so that they don't compete with
the parent's internal process. This patch broke this rule by enabling
the controller on the leaf node by default. Let's revert to the original
approach of not enabling the controller on the leaf cgroup node.

Signed-by-off: Kamalesh Babulal <kamalesh.babulal@oracle.com>