Revert "api: cgroup v2 – read subtree_control of the leaf node"
This reverts commit 39012eeefa454e09fc4192cacddb508ee00abee6.

As per the no internal process constraint, of cgroup v2 no controller
should be enabled in the leaf cgroup node. This patch breaks this
constraint by assuming the controllers are enabled until the leaf node
of the hierarchy. Let's revert to the original approach of reading
until the parent of the leaf cgroup node.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>