api.c: add cgroup v2 support to cgroup_get_current_controller_path()
Add support to read/parse cgroup v2 format /proc/<pid>/cgroup to
cgroup_get_current_controller_path(). The format of /proc/<pid>/cgroup
remains the same:
hierarchy-ID:controller-list:cgroup-path

the difference is that in cgroup v2 mode:
- hierarchy-ID is always 0 (one hierarchy allowed)
- controller-list is empty

there can also be a mix of both legacy and unified controllers for a
given task if the cgroup mode is hybrid mode.  This patch, add support
for the mixed cgroup setup mode too.

Reported-by: Adriaan Schmidt adriaan.schmidt@siemens.com
Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>