tools/cgcreate: add -p <pid> support for systemd scope
Currently, an idle_thread is created for the transient systemd scope by
default, it works fine, when the slice/scope creation is treated in par
to create of traditional cgroup, but the user may wish to create a
slice/scope, with a process of their choice, instead of a default idle
thread.  Add support to pass task pid via '-p/--pid' option, while
creating a systemd slice/scope.

 # ./src/tools/cgcreate -c -gcpu:foo.slice/bar.scope --pid=4253

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>