api: Delete the cgroup if cgroup_create_cgroup() fails
cgroup_create_cgroup() loops through the controllers in struct cgroup
and creates the requested cgroups.  Some controllers could succeed
before a controller fails.  This results in a case where the cgroup
directory has been created and remains in /sys/fs/cgroup even after a
failure.

Invoke cgroup_delete_cgroup() after we fail to create a cgroup under a
controller.  This will delete the cgroup directory if it had been
created and put the system (nearly) back to its original state.  (It's
possible that the parent cgroup's subtree_control file was modified and
not restored, but that's best fixed in another patch.)

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Reviewed-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>