api: Support NULL controllers[] in cgroup_change_cgroup_path()
cgroup_change_cgroup_path() allows the user to move a pid to a
different cgroup (via the cgroup name and not a cgroup struct).
In cgroup v1, it's imperative that the controller(s) are
provided, so that the proper path can be built up.  In cgroup v2,
the list of controllers is optional, since they are not a part
of the path.

Add support for the controllers[] list to be NULL iff we are
running in unified (cgroup v2) mode.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Reviewed-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>