config: systemd: Add a return code to cgroup_set_default_systemd_cgroup()
Add a return code to cgroup_set_default_systemd_cgroup().  This is
critical for users of the libcgroup C API, as they need to be certain
that a default slice/scope is set.  If they mistakenly think one is
set when it is not, then they could erroneously operate on the root
cgroup.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>