tools/cgexec: add -r option for systemd scope
Provide an option for users to replace the default idle_thread created
for systemd scope.  With '-r' option specified, the program passed as
the argument to cgexec will replace the default idle_thread, if it
exists.  Also, if it's the only task, killing it will remove the .scope
cgroup it's running in.

$ sudo cgexec -r -gcpu:libcgroup.slice/db.scope ./new_default

For non scope cgroups, it launches the task program. The algorithm in
brief is as follows:
Once the current task, instance of cgexec is migrated to the
expected cgroup(s):
1. fork() and in the child path perform the exec to program passed in
   the command line.  This means, we have cgexec instance and program
   executing in the desired cgroup(s).
2. In the parent, parse /proc/<pid>/cgroup of the current process
   (cgexec) and if the cgroup name ends with .scope and has a task
   with /proc/<pid>/cmdline as "libcgroup_systemd_idle_thread", kill
   the task. This gives the illusion of replacement.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>