api: Show all members of a given group
In the context of a group rule (i.e. indicated by '@' used to prefix the
actual group name), getgrnam(3) is used to provide a pointer to a group
file entry that may contain a NULL-terminated array of pointers to group
members. A user can belong to multiple groups. With this information, we
then check the username that corresponds to the specified UID against
each group member for a match. This patch makes it possible to see this
information if debug level logging is enabled.

For example, this can be useful when troubleshooting the Cgroup rules
engine daemon.

Signed-off-by: Aaron Tomlin <atomlin@atomlin.com>