ftests: Fix test 005 on Cgroup v2 Ubuntu22
Test 005-cgsnapshot-basic_snapshot_v2.py was failing on Ubuntu22
(non-sudo tests) because the cpuset.cpus.exclusive.effective field
is no longer being reported by the kernel in that case.

Remove that line from the expected output for the non-sudo tests but
leave it in for the sudo tests as it's still reported there.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>