ftests: Retry getting pid of sleep process
The ftests framework uses sleeping processes to test various cgroup
features.  Sometimes ps throws an error when we try to enumerate
processes via `ps x` within an lxc container.

The error thrown by ps is as follows:
	Signal 21 (TTIN) caught by ps (3.3.17).
	ps:ps/display.c:70: please report this bug

Fix this by adding a retry loop for the ps command, as it usually works
in a subsequent pass.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Acked-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>