tests: fix checking wrong result in 017-API_fuzz_test
There seem to be various calls to cgroup_add_controller where the result
isn't validated, instead checking the cgrp pointer.

We could add two asserts in each place being changed here if the concern
was testing for if cgroup_add_controller mangled cgrp, but it looks
more likely that it's a copy/paste error instead and we're just testing
the wrong thing.

Signed-off-by: Sam James <sam@gentoo.org>