ftests: Fix false positive in 010 testcase
Fix the false positive in the testcase number 010, the return value
should be by default `TEST_FAILED`, instead of `TEST_PASSED`. The flow
of the testcase is designed to set the return value to `TEST_PASSED`
only if the current kernel's cpu controller settings matches any of
the expected cpu controller output list.

Fixes: 88ffe562376f ("ftests/010: Refactor code to match outputs with same line")
Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>