ftests/utils: Move get_distro() to distro/consts_distro.py
get_distro() helper was introduced to choose the right cpu controller
settings list, based on the underlying distribution. Now that, the
consts are moved to the consts directory, move the helper also to the
distro/consts_distro.py file. Also, add the logic to raise an Expection
when unsupported distro found.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>